# docs/_archivo-v1/02-arquitectura/2.4-contratos/openapi.yaml
openapi: 3.1.0
info:
  title: Acalud API
  version: 0.1.0
  description: >
    Contrato de la API del monolito modular (ADR-002). Fuente de verdad para los
    contract tests (pirámide ADR-002). Cada operación referencia su CU (1.1).
    Errores en RFC 9457 (application/problem+json). Campos JSON en snake_case español
    (Glosario 0.2). Autenticación dual (ADR-004): cookie httpOnly (web) o Bearer (APK).
servers:
  - url: /api/v1
security:
  - cookieAuth: []
  - bearerAuth: []
...
paths:
...
  /instituciones:
    post:
      tags: [instituciones]
      summary: Registrar institución (CU-023; crea membresía encargado en la misma transacción)
      requestBody: { content: { application/json: { schema: { $ref: '#/components/schemas/InstitucionInput' } } } }
      responses:
        '201': { description: Institución activa (S-11), content: { application/json: { schema: { type: object, properties: { institucion_id: { type: string, format: uuid } } } } } }
        '409': { description: 'CU-023 A1/A2/A3: usuario ya vinculado, CUIT o email institucional duplicado' }
        '422': { $ref: '#/components/responses/Problem422' }
  /instituciones/{institucion_id}/asignaciones:
    post:
      tags: [instituciones]
      summary: Asignar licencias a docentes (CU-026; transacción única + notificación dashboard/email + auditoría)
      parameters: [ { $ref: '#/components/parameters/institucion_id' } ]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [producto_id, asignaciones]
              properties:
                producto_id: { type: string, format: uuid }
                asignaciones:
                  type: array
                  minItems: 1
                  items:
                    type: object
                    required: [docente_id, cantidad]
                    properties:
                      docente_id: { type: string, format: uuid }
                      cantidad: { type: integer, minimum: 1 }
                observaciones: { type: [string, 'null'], maxLength: 500 }
      responses:
        '201': { description: 'Asignadas; devuelve ids de asignación y disponibilidad actualizada' }
        '404': { description: 'Sin permisos de encargado (RN-001) o producto fuera del inventario' }
        '422': { description: 'Docente no vinculado (A3) o licencias insuficientes (A1, con disponible)' }
  /instituciones/{institucion_id}/revocaciones:
    post:
      tags: [instituciones]
      summary: Revocar licencias (CU-027; parcial A9 o total A10 — la fila NUNCA se borra, pasa a revoked)
      parameters: [ { $ref: '#/components/parameters/institucion_id' } ]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [docente_id, producto_id, cantidad_a_revocar]
              properties:
                docente_id: { type: string, format: uuid }
                producto_id: { type: string, format: uuid }
                cantidad_a_revocar: { type: integer, minimum: 1 }
                observaciones: { type: [string, 'null'], maxLength: 500, description: 'Razón de la revocación (revocation_reason)' }
      responses:
        '200': { description: 'Revocadas; devuelve cantidad_revocada y cantidad_restante' }
        '404': { description: 'Sin permisos (RN-001), producto fuera del inventario o docente sin asignación activa (A3)' }
        '422': { description: 'Cantidad inválida (A1) o superior a la asignada (A2, con la actual)' }
  /instituciones/{institucion_id}/inventario:
    get:
      tags: [instituciones]
      summary: Inventario institucional con resumen (CU-025; solo encargado is_admin; registra inventory_viewed)
      parameters:
        - { $ref: '#/components/parameters/institucion_id' }
        - { name: producto_id, in: query, schema: { type: string, format: uuid }, description: 'Filtro A5' }
        - { name: orden, in: query, schema: { type: string, enum: [cantidad_adquirida, cantidad_asignada, cantidad_disponible, ultima_compra, nombre] }, description: 'Orden A6' }
        - { name: direccion, in: query, schema: { type: string, enum: [asc, desc] } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/InventarioView' } } } }
        '404': { description: 'Institución inexistente o no visible para el caller (CU-025 A2; recurso ajeno → 404)' }
  /instituciones/{institucion_id}/inventario/{producto_id}:
    get:
      tags: [instituciones]
      summary: Detalle de producto del inventario con historial de compras y docentes asignados (CU-025 A7)
      parameters:
        - { $ref: '#/components/parameters/institucion_id' }
        - { name: producto_id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/InventarioDetalleView' } } } }
        '404': { description: 'Sin permisos (A2) o producto fuera del inventario (A7)' }
...
  /instituciones/{institucion_id}/docentes/asignaciones:
    get:
      tags: [instituciones]
      summary: Listado de docentes vinculados con sus asignaciones (CU-028; A6 filtro producto, A7 búsqueda, A8 orden; registra assigned_teachers_viewed)
      parameters:
        - { $ref: '#/components/parameters/institucion_id' }
        - { name: producto_id, in: query, schema: { type: string, format: uuid }, description: 'Filtro A6' }
        - { name: buscar, in: query, schema: { type: string, maxLength: 200 }, description: 'Búsqueda por nombre A7' }
        - { name: orden, in: query, schema: { type: string, enum: [total_licencias, nombre] }, description: 'Orden A8' }
        - { name: direccion, in: query, schema: { type: string, enum: [asc, desc] } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/DocentesAsignadosView' } } } }
        '404': { description: 'Sin permisos de encargado (RN-001)' }
  /instituciones/{institucion_id}/docentes/asignaciones/{docente_id}:
    get:
      tags: [instituciones]
      summary: Detalle de asignaciones de un docente, incluidas las revocadas (CU-028 A9; RN-007/RN-008)
      parameters:
        - { $ref: '#/components/parameters/institucion_id' }
        - { name: docente_id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/DocenteDetalleAsignacionesView' } } } }
        '404': { description: 'Sin permisos (RN-001) o docente no vinculado a la institución (A9)' }
  /instituciones/{institucion_id}/docentes:
    get:
      tags: [instituciones]
      summary: Listado de membresías (CU-027; desvinculadas bajo filtro)
      parameters: [ { $ref: '#/components/parameters/institucion_id' }, { name: estado, in: query, schema: { type: string, enum: [activa, invitada, desvinculada, vencida] } } ]
      responses: { '200': { description: OK, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/MembresiaView' } } } } } }
...
components:
  securitySchemes:
    cookieAuth: { type: apiKey, in: cookie, name: acalud_sesion }
    bearerAuth: { type: http, scheme: bearer }
  parameters:
...
    institucion_id: { name: institucion_id, in: path, required: true, schema: { type: string, format: uuid } }
...
  responses:
...
    Problem422: { description: Validación de negocio, content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } } }
...
  schemas:
    Problem:
      type: object
      description: RFC 9457
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        trace_id: { type: string }
...
    Domicilio:
      type: object
      required: [calle, numero, codigo_postal, provincia, localidad]
      properties:
        calle: { type: string }
        numero: { type: string }
        codigo_postal: { type: string }
        provincia: { type: string }
        localidad: { type: string }
...
    InstitucionInput:
      type: object
      required: [nombre_legal, identificador_tributario, email_institucional, domicilio]
      properties:
        nombre_legal: { type: string, maxLength: 200 }
        identificador_tributario: { type: string, description: 'CUIT con o sin guiones; se persiste normalizado a 11 dígitos (CU-023 RN-001)' }
        email_institucional: { type: string, format: email, maxLength: 255 }
        domicilio: { $ref: '#/components/schemas/Domicilio' }
        telefono: { type: [string, 'null'], maxLength: 30 }
        nivel_educativo: { type: [string, 'null'], description: 'Nombre de una fila de levels (Inicial, Primaria, Secundaria; CU-023 A11)' }
        cantidad_alumnos: { type: [integer, 'null'], minimum: 1 }
    InventarioItemView:
      type: object
      properties:
        producto_id: { type: string, format: uuid }
        nombre_producto: { type: string }
        descripcion_producto: { type: [string, 'null'] }
        cantidad_adquirida: { type: integer }
        cantidad_asignada: { type: integer }
        cantidad_disponible: { type: integer, description: 'adquirida − asignada (CU-025 RN-002)' }
        ultima_compra_en: { type: [string, 'null'], format: date-time }
        total_gastado: { type: [number, 'null'], description: 'Suma de compras b2b pagas (CU-025 p6, opcional)' }
    InventarioView:
      type: object
      properties:
        institucion_id: { type: string, format: uuid }
        resumen:
          type: object
          properties:
            total_adquiridas: { type: integer }
            docentes_asignados: { type: integer }
            total_en_uso: { type: integer }
            total_disponibles: { type: integer }
        items: { type: array, items: { $ref: '#/components/schemas/InventarioItemView' } }
    InventarioDetalleView:
      type: object
      properties:
        producto_id: { type: string, format: uuid }
        nombre_producto: { type: string }
        descripcion_producto: { type: [string, 'null'] }
        precio: { type: number }
        cantidad_adquirida: { type: integer }
        cantidad_asignada: { type: integer }
        cantidad_disponible: { type: integer }
        compras:
          type: array
          items:
            type: object
            properties:
              orden_id: { type: string, format: uuid }
              numero: { type: string }
              fecha: { type: string, format: date-time }
              cantidad: { type: integer }
              monto: { type: number }
        docentes:
          type: array
          items:
            type: object
            properties:
              docente_id: { type: string, format: uuid }
              nombre: { type: string }
              cantidad: { type: integer }
              asignada_en: { type: string, format: date-time }
              estado: { type: string, enum: [active, revoked] }
...
    AsignacionDeDocenteView:
      type: object
      properties:
        producto_id: { type: string, format: uuid }
        nombre_producto: { type: string }
        cantidad: { type: integer }
        asignada_en: { type: string, format: date-time }
        asignada_por: { type: [string, 'null'] }
        estado: { type: string, enum: [active, revoked] }
    DocentesAsignadosView:
      type: object
      properties:
        institucion_id: { type: string, format: uuid }
        resumen:
          type: object
          properties:
            total_docentes_con_asignaciones: { type: integer }
            total_licencias_asignadas: { type: integer }
            productos_mas_asignados:
              type: array
              items:
                type: object
                properties:
                  producto_id: { type: string, format: uuid }
                  nombre_producto: { type: string }
                  total: { type: integer }
        docentes:
          type: array
          items:
            type: object
            properties:
              docente_id: { type: string, format: uuid }
              nombre: { type: string }
              email: { type: string }
              total_licencias: { type: integer }
              ultima_asignacion_en: { type: [string, 'null'], format: date-time }
              asignaciones: { type: array, items: { $ref: '#/components/schemas/AsignacionDeDocenteView' } }
    DocenteDetalleAsignacionesView:
      type: object
      properties:
        docente_id: { type: string, format: uuid }
        nombre: { type: string }
        email: { type: string }
        vinculado_en: { type: [string, 'null'], format: date-time }
        asignaciones:
          type: array
          items:
            allOf:
              - $ref: '#/components/schemas/AsignacionDeDocenteView'
              - type: object
                properties:
                  revocada_en: { type: [string, 'null'], format: date-time }
                  revocada_por: { type: [string, 'null'] }
                  razon_revocacion: { type: [string, 'null'] }
    MembresiaView:
      type: object
      properties:
        membresia_id: { type: string, format: uuid }
        nombre: { type: [string, 'null'] }
        email: { type: string }
        rol: { type: string }
        estado: { type: string }
        fecha_alta: { type: [string, 'null'], format: date }
        sesiones_cargadas: { type: integer }
...